Move repositories into tests.
Removes a bit of the clutter in the test packages.
