+++ spring-data-neo4j-rx/src/test/java/org/neo4j/springframework/data/integration/imperative/ProjectionIT.java
@@ -33,6 +33,7 @@ import org.neo4j.springframework.data.integration.shared.NamesOnly;
 import org.neo4j.springframework.data.integration.shared.NamesOnlyDto;
 import org.neo4j.springframework.data.integration.shared.Person;
 import org.neo4j.springframework.data.integration.shared.PersonSummary;
+import org.neo4j.springframework.data.repository.Neo4jRepository;
 import org.neo4j.springframework.data.repository.config.EnableNeo4jRepositories;
 import org.neo4j.springframework.data.test.Neo4jExtension;
 import org.neo4j.springframework.data.test.Neo4jIntegrationTest;
@@ -45,7 +46,7 @@ import org.springframework.transaction.annotation.EnableTransactionManagement;
  * @author Gerrit Meier
  */
 @Neo4jIntegrationTest
-public class ProjectionIT {
+class ProjectionIT {
 
 	private static final String FIRST_NAME = "Hans";
 	private static final String LAST_NAME = "Mueller";
@@ -152,8 +153,19 @@ public class ProjectionIT {
 
 	}
 
+	public interface ProjectionPersonRepository extends Neo4jRepository<Person, Long> {
+
+		Collection<NamesOnly> findByLastName(String lastName);
+
+		Collection<PersonSummary> findByFirstName(String firstName);
+
+		Collection<NamesOnlyDto> findByFirstNameAndLastName(String firstName, String lastName);
+
+		<T> Collection<T> findByLastNameAndFirstName(String lastName, String firstName, Class<T> projectionClass);
+	}
+
 	@Configuration
-	@EnableNeo4jRepositories
+	@EnableNeo4jRepositories(considerNestedRepositories = true)
 	@EnableTransactionManagement
 	static class Config extends AbstractNeo4jConfig {
